Initial Commit
`npm create tauri-app@latest`
https://github.com/borewit/music-metadata

My goal here is to make an iTunes/Music app alternative, which is built with SolidJS, Tauri, and SQLite for the metadata database. The audio files themselves would still be saved to disk in the music folder like iTunes/Music does currently.

I also want to add proper metadata editing support as that isn't as easy to do as it really should be I feel.

I also want to have just an easy way to have a music library, and not be tied to a specific developer or platform, be it if it's closed source, proprietary, or needs a subscription to use it. I am personally very sick of how splintered the music scene is in the tech world as it currently stands. I feel like I see the next wave of how personal music can get again, and I think it will come to that point, but I don't think we're there yet. Places like Bandcamp are a great example of where we are going. Same with lossless formats. All of the tech for a great experience is all there, it's just that there isn't really a streamlined way for music management yet, which is bumming to me. It seems only natural for the web to be a place to help foster that growth. I want to go back to the days of record and CD collections, but in a digital format. Taking pride in your favorite music should be an easy thing, you support the artists, you have your copies, and you can easily catalog and share the info around those tracks. The info part is what feels missing to me, it is so messy in terms of file formats and things of that nature, that's what I'm hoping to streamline. Same with the usage of ffmpeg and multimedia formats, having that as an open GUI seems only logical as well, all of the tech for these plans are all already out there, it just hasn't come to fruition just yet. If these tools weren't available to make all of this happen, this undertaking would be very challenging. However, we are not at that point. It's just a matter of making things more user-accessible and future-proof. Standardization :)

'And The Hero Will Drown' by Story Of The Year kind of sounds like New Noise. Listening to the Need for Speed Underground soundtrack. Listened to "Instrumental Math Rock Mix" before this.